// vrm/vrm-interfaces.ts
export interface IVRMBlendShapeMaterialBind {
  materialName: string;
  propertyName: string;
  targetValue: number[];
}